+++ src/config/configStore.ts
@@ -6,8 +6,16 @@
  */
 
 import { AsyncCreatable, set } from '@salesforce/kit';
-import { AnyJson, definiteEntriesOf, definiteValuesOf, get, getAnyJson, JsonMap, Optional } from '@salesforce/ts-types';
-import { Dictionary } from '@salesforce/ts-types';
+import {
+  AnyJson,
+  definiteEntriesOf,
+  definiteValuesOf,
+  Dictionary,
+  get,
+  getAnyJson,
+  JsonMap,
+  Optional
+} from '@salesforce/ts-types';
 
 /**
  * The allowed types stored in a config store.
@@ -69,7 +77,7 @@ export abstract class BaseConfigStore<T extends BaseConfigStore.Options> extends
   public constructor(options: T) {
     super(options);
     this.options = options;
-    this.setContents(this.options.contents || {});
+    this.setContents(this.options.contents);
   }
 
   /**
@@ -216,7 +224,7 @@ export abstract class BaseConfigStore<T extends BaseConfigStore.Options> extends
   }
 
   // Allows extended classes the ability to override the set method. i.e. maybe they don't want
-  // nexted object set from kit.
+  // nested object set from kit.
   protected setMethod(contents: ConfigContents, key: string, value?: ConfigValue) {
     set(contents, key, value);
   }