+++ src/schema/validator.ts
@@ -38,7 +38,7 @@ export class SchemaValidator {
    * Creates a new `SchemaValidator` instance given a logger and path to a schema file.
    *
    * @param logger An {@link Logger} instance on which to base this class's logger.
-   * @param schemaPath The path from which the schema with which to validate should be loaded.
+   * @param schemaPath The path to the schema file to load and use for validation.
    */
   public constructor(logger: Logger, private schemaPath: string) {
     this.logger = logger.child('SchemaValidator');