+++ src/util/fs.ts
@@ -7,11 +7,14 @@
 
 import { parseJson, parseJsonMap } from '@salesforce/kit';
 import { AnyJson, JsonMap, Optional } from '@salesforce/ts-types';
+import * as crypto from 'crypto';
 import * as fsLib from 'fs';
 import * as mkdirpLib from 'mkdirp';
 import * as path from 'path';
 import { promisify } from 'util';
 import { SfdxError } from '../sfdxError';
+
+type PerformFunction = (filePath: string, file?: string, dir?: string) => Promise<void>;
 
 export const fs = {
   /**
@@ -74,7 +77,7 @@ export const fs = {
    * Promisified version of {@link https://npmjs.com/package/mkdirp|mkdirp}.
    */
   // @ts-ignore TODO: figure out how to bind to correct promisify overload
-  mkdirp: (folderPath: string, mode?: string | object): Promise<void> => promisify(mkdirpLib)(folderPath, mode),
+  mkdirp: (folderPath: string, mode?: string | object): Promise<void> => mkdirpLib(folderPath, mode),
 
   /**
    * Deletes a folder recursively, removing all descending files and folders.
@@ -158,5 +161,82 @@ export const fs = {
       encoding: 'utf8',
       mode: fs.DEFAULT_USER_FILE_MODE
     });
+  },
+
+  /**
+   * Checks if a file path exists
+   *
+   * @param filePath the file path to check the existence of
+   */
+  fileExists: async (filePath: string): Promise<boolean> => {
+    try {
+      await fs.access(filePath);
+      return true;
+    } catch (err) {
+      return false;
+    }
+  },
+
+  /**
+   * Recursively act on all files or directories in a directory
+   *
+   * @param dir path to directory
+   * @param perform function to be run on contents of dir
+   * @param onType optional parameter to specify type to actOn
+   * @returns void
+   */
+
+  actOn: async (dir: string, perform: PerformFunction, onType: 'file' | 'dir' = 'file'): Promise<void> => {
+    for (const file of await fs.readdir(dir)) {
+      const filePath = path.join(dir, file);
+      const stat = await fs.stat(filePath);
+
+      if (stat) {
+        if (stat.isDirectory()) {
+          await fs.actOn(filePath, perform, onType);
+          if (onType === 'dir') {
+            await perform(filePath);
+          }
+        } else if (stat.isFile() && onType === 'file') {
+          await perform(filePath, file, dir);
+        }
+      }
+    }
+  },
+
+  /**
+   * Checks if files are the same
+   *
+   * @param file1Path the first file path to check
+   * @param file2Path the second file path to check
+   * @returns boolean
+   */
+  areFilesEqual: async (file1Path: string, file2Path: string): Promise<boolean> => {
+    try {
+      const file1Size = (await fs.stat(file1Path)).size;
+      const file2Size = (await fs.stat(file2Path)).size;
+      if (file1Size !== file2Size) {
+        return false;
+      }
+
+      const contentA = await fs.readFile(file1Path);
+      const contentB = await fs.readFile(file2Path);
+
+      return fs.getContentHash(contentA) === fs.getContentHash(contentB);
+    } catch (err) {
+      throw new SfdxError(`The path: ${err.path} doesn't exist or access is denied.`, 'DirMissingOrNoAccess');
+    }
+  },
+
+  /**
+   * Creates a hash for the string that's passed in
+   * @param contents The string passed into the function
+   * @returns string
+   */
+  getContentHash(contents: string | Buffer) {
+    return crypto
+      .createHash('sha1')
+      .update(contents)
+      .digest('hex');
   }
 };