+++ test/unit/authInfoTest.ts
@@ -13,6 +13,7 @@ import { OAuth2, OAuth2Options } from 'jsforce';
 // @ts-ignore WebStorm is reporting an error for the nested import
 import * as Transport from 'jsforce/lib/transport';
 import * as jwt from 'jsonwebtoken';
+import * as pathImport from 'path';
 import { AuthFields, AuthInfo, OAuth2WithVerifier } from '../../src/authInfo';
 import { AuthInfoConfig } from '../../src/config/authInfoConfig';
 import { ConfigFile } from '../../src/config/configFile';
@@ -1333,6 +1334,22 @@ describe('AuthInfo', () => {
       expect(testCallback.firstCall.args[1]).to.equal(testMetadata.accessToken);
     });
 
+    it('should path.resolve jwtkeyfilepath', async () => {
+      const pathSpy = $$.SANDBOX.spy(pathImport, 'resolve');
+      const context = {
+        update: () => {},
+        buildJwtConfig: () => {},
+        isTokenOptions: () => false,
+        getUsername: () => '',
+        privateKeyFile: 'authInfoTest/jwt/server.key',
+        options: {}
+      };
+
+      await AuthInfo.prototype['initAuthOptions'].call(context, context, null);
+
+      expect(pathSpy.calledOnce).to.be.true;
+    });
+
     it('should call the callback with OrgDataNotAvailableError when AuthInfo.init() fails', async () => {
       const context = {
         getUsername: () => '',
@@ -1434,6 +1451,10 @@ describe('AuthInfo', () => {
 
   describe('audienceUrl', () => {
     const sfdxAudienceUrlSetting = process.env.SFDX_AUDIENCE_URL;
+
+    beforeEach(() => {
+      process.env.SFDX_AUDIENCE_URL = '';
+    });
 
     afterEach(() => {
       process.env.SFDX_AUDIENCE_URL = sfdxAudienceUrlSetting || '';